update ndpi: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@59116 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/ndpi-netfilter/src/include/ndpi_replace_printf.h
@@ -55,7 +55,7 @@
 
 #endif
 
-void ndpi_debug_printf(unsigned int proto, struct ndpi_detection_module_struct *ndpi_str, ndpi_log_level_t log_level,
+void ndpi_debug_printf(u_int16_t proto, struct ndpi_detection_module_struct *ndpi_str, ndpi_log_level_t log_level,
                        const char *file_name, const char *func_name, int line_number, const char *format, ...);
 
 #endif
